requirements.txt
streamlit
pandas
python-docx
PyMuPDF
transformers
torch
scikit-learn
nltk
spacy
sentence-transformers
mysql-connector-python